refactor(skills): restructure skill types and update related components
