+++ src/modules/skills/skills.const.ts
@@ -1,16 +1,4 @@
-import type { SkillDifficulty, SkillStatus } from "./skills.types";
-
-// import img1 from "../../assets/skills/1.png";
-// import img2 from "../../assets/skills/2.png";
-// import img3 from "../../assets/skills/3.png";
-// import img4 from "../../assets/skills/4.png";
-// import img5 from "../../assets/skills/5.png";
-// import img6 from "../../assets/skills/6.png";
-// import img7 from "../../assets/skills/7.png";
-// import img8 from "../../assets/skills/8.png";
-// import img9 from "../../assets/skills/9.png";
-// import img10 from "../../assets/skills/10.png";
-// import img11 from "../../assets/skills/11.png";
+import type { FilterSkillDifficulty, FilterSkillStatus } from "./skills.types";
 
 export const difficultyColors = {
   EASY: "bg-green-100 text-green-800",
@@ -65,22 +53,22 @@ export const difficultyDetailLabels = {
 export const SKILLS_PER_PAGE = 12;
 
 // Fonctions utilitaires pour les couleurs et labels
-export const getDifficultyColor = (difficulty: SkillDifficulty) => {
+export const getDifficultyColor = (difficulty: FilterSkillDifficulty) => {
   if (difficulty === "ALL") return "bg-gray-500";
   return difficultyDetailColors[difficulty as keyof typeof difficultyDetailColors] || "bg-gray-500";
 };
 
-export const getStatusColor = (status: SkillStatus) => {
+export const getStatusColor = (status: FilterSkillStatus) => {
   if (status === "ALL") return "bg-gray-500";
   return statusDetailColors[status as keyof typeof statusDetailColors] || "bg-gray-500";
 };
 
-export const getStatusLabel = (status: SkillStatus) => {
+export const getStatusLabel = (status: FilterSkillStatus) => {
   if (status === "ALL") return status;
   return statusDetailLabels[status as keyof typeof statusDetailLabels] || status;
 };
 
-export const getDifficultyLabel = (difficulty: SkillDifficulty) => {
+export const getDifficultyLabel = (difficulty: FilterSkillDifficulty) => {
   if (difficulty === "ALL") return difficulty;
   return difficultyDetailLabels[difficulty as keyof typeof difficultyDetailLabels] || difficulty;
 };
