+++ src/modules/stats/components/chart/index.ts
@@ -1,8 +1,5 @@
 // Types
-export type {
-  ChartData,
-  UserStats,
-} from "../../types/stats.types";
+export type { ChartData, UserStats } from "../../types/stats.types";
 
 export { SkillExperienceChart } from "./skill-experience-chart/SkillExperienceChart";
