+++ src/shared/services/skill/api-skill.type.ts
@@ -4,10 +4,6 @@ export type CreateSkillInput = Skill;
 
 export type CreateSkillResponse = Skill;
 
-export type UpdateSkillInput = Partial<Skill> & {
-  totalXp: number;
-  completedQuests: number;
-  averageQuestXp: number;
-};
+export type UpdateSkillInput = Partial<Skill>;
 
 export type UpdateSkillResponse = Skill;
